+++ client/tsconfig.json
@@ -35,7 +35,7 @@
     "src/**/*.vue",
     "tests/**/*.ts",
     "tests/**/*.tsx"
-  ],
+, "src/api/api.js"  ],
   "exclude": [
     "node_modules"
   ]
